+++ tests/test_needuml.py
@@ -202,3 +202,191 @@ def test_needuml_diagram_allowmixing(test_app):
 
     out = subprocess.run(["sphinx-build", "-M", "html", srcdir, out_dir], capture_output=True)
     assert out.returncode == 0
+
+
+@pytest.mark.parametrize("test_app", [{"buildername": "html", "srcdir": "doc_test/doc_needuml_save"}], indirect=True)
+def test_needuml_save(test_app):
+    app = test_app
+    app.build()
+
+    # check generated plantuml code saved in given path
+    from sys import platform
+
+    if platform == "win32":
+        assert "doc_needuml_save\\_build\\html" in str(app.outdir)
+    else:
+        assert "doc_needuml_save/_build/html" in str(app.outdir)
+    assert app.config.needs_build_needumls == "my_needumls"
+    saved_uml_path_01 = Path(app.outdir) / app.config.needs_build_needumls / "_build/my_needuml.puml"
+    assert saved_uml_path_01.exists()
+
+    with open(saved_uml_path_01) as f1:
+        f1_contents = f1.readlines()
+        assert len(f1_contents) == 5
+        assert f1_contents[0] == "@startuml\n"
+        assert f1_contents[1] == "\n"
+        assert f1_contents[2] == "DC -> Marvel: Hi Kevin\n"
+        assert f1_contents[3] == "Marvel --> DC: Anyone there?\n"
+        assert f1_contents[4] == "@enduml\n"
+
+    saved_uml_path_02 = Path(app.outdir) / app.config.needs_build_needumls / "_out/sub_folder/my_needs.puml"
+    assert saved_uml_path_02.exists()
+
+    with open(saved_uml_path_02) as f2:
+        f2_contents = f2.readlines()
+
+        assert len(f2_contents) == 7
+
+        assert f2_contents[0] == "@startuml\n"
+        assert f2_contents[1] == "\n"
+
+        assert "User Story" in f2_contents[2]
+        assert "Test story" in f2_contents[2]
+        assert "ST_001" in f2_contents[2]
+
+        assert f2_contents[3] == "\n"
+
+        assert "User Story" in f2_contents[4]
+        assert "Test story 2" in f2_contents[4]
+        assert "ST_002" in f2_contents[4]
+
+        assert f2_contents[5] == "\n"
+        assert f2_contents[6] == "@enduml\n"
+
+
+@pytest.mark.parametrize(
+    "test_app", [{"buildername": "html", "srcdir": "doc_test/doc_needuml_save_with_abs_path"}], indirect=True
+)
+def test_needuml_save_with_abs_path(test_app):
+    import subprocess
+
+    app = test_app
+
+    srcdir = Path(app.srcdir)
+    out_dir = srcdir / "_build"
+
+    out = subprocess.run(["sphinx-build", "-M", "html", srcdir, out_dir], capture_output=True)
+    assert out.returncode == 1
+
+    assert (
+        "sphinx_needs.directives.needuml.NeedumlException: "
+        "Given save path: /_out/my_needuml.puml, is not a relative path." in out.stderr.decode("utf-8")
+    )
+
+
+@pytest.mark.parametrize(
+    "test_app", [{"buildername": "needumls", "srcdir": "doc_test/doc_needuml_save"}], indirect=True
+)
+def test_needumls_builder(test_app):
+    app = test_app
+    app.build()
+
+    # check generated plantuml code saved in given path
+    from sys import platform
+
+    if platform == "win32":
+        assert "doc_needuml_save\\_build\\needumls" in str(app.outdir)
+    else:
+        assert "doc_needuml_save/_build/needumls" in str(app.outdir)
+    saved_uml_path_01 = Path(app.outdir) / "_build/my_needuml.puml"
+    assert saved_uml_path_01.exists()
+
+    with open(saved_uml_path_01) as f1:
+        f1_contents = f1.readlines()
+        assert len(f1_contents) == 5
+        assert f1_contents[0] == "@startuml\n"
+        assert f1_contents[1] == "\n"
+        assert f1_contents[2] == "DC -> Marvel: Hi Kevin\n"
+        assert f1_contents[3] == "Marvel --> DC: Anyone there?\n"
+        assert f1_contents[4] == "@enduml\n"
+
+    saved_uml_path_02 = Path(app.outdir) / "_out/sub_folder/my_needs.puml"
+    assert saved_uml_path_02.exists()
+
+    with open(saved_uml_path_02) as f2:
+        f2_contents = f2.readlines()
+
+        assert len(f2_contents) == 7
+
+        assert f2_contents[0] == "@startuml\n"
+        assert f2_contents[1] == "\n"
+
+        assert "User Story" in f2_contents[2]
+        assert "Test story" in f2_contents[2]
+        assert "ST_001" in f2_contents[2]
+
+        assert f2_contents[3] == "\n"
+
+        assert "User Story" in f2_contents[4]
+        assert "Test story 2" in f2_contents[4]
+        assert "ST_002" in f2_contents[4]
+
+        assert f2_contents[5] == "\n"
+        assert f2_contents[6] == "@enduml\n"
+
+
+@pytest.mark.parametrize("test_app", [{"buildername": "html", "srcdir": "doc_test/doc_needuml_filter"}], indirect=True)
+def test_needuml_filter(test_app):
+    app = test_app
+    app.build()
+
+    all_needumls = app.env.needs_all_needumls
+    assert len(all_needumls) == 1
+
+    needuml_01_content = all_needumls["needuml-index-0"]["content"]
+    assert "DC -> Marvel: Hi Kevin\n" in needuml_01_content
+    assert "Marvel --> DC: Anyone there?\n\n" in needuml_01_content
+    assert "{% for need in filter(\"type == 'story' and status != 'open'\") %}" in needuml_01_content
+
+    html = Path(app.outdir, "index.html").read_text(encoding="utf8")
+    assert "as ST_002 [[../index.html#ST_002]]" in html
+
+    import subprocess
+
+    srcdir = Path(app.srcdir)
+    out_dir = srcdir / "_build"
+
+    out = subprocess.run(["sphinx-build", "-M", "html", srcdir, out_dir], capture_output=True)
+    assert out.returncode == 0
+
+
+@pytest.mark.parametrize(
+    "test_app", [{"buildername": "html", "srcdir": "doc_test/doc_needuml_jinja_func_flow"}], indirect=True
+)
+def test_needuml_jinja_func_flow(test_app):
+    app = test_app
+    app.build()
+
+    all_needumls = app.env.needs_all_needumls
+    assert len(all_needumls) == 1
+
+    assert '{{flow("ST_001")}}' in all_needumls["needuml-index-0"]["content"]
+
+    html = Path(app.outdir, "index.html").read_text(encoding="utf8")
+    assert "as ST_001 [[../index.html#ST_001]]" in html
+
+    import subprocess
+
+    srcdir = Path(app.srcdir)
+    out_dir = srcdir / "_build"
+
+    out = subprocess.run(["sphinx-build", "-M", "html", srcdir, out_dir], capture_output=True)
+    assert out.returncode == 0
+
+
+@pytest.mark.parametrize(
+    "test_app", [{"buildername": "html", "srcdir": "doc_test/doc_needuml_jinja_func_need_removed"}], indirect=True
+)
+def test_needuml_jinja_func_need_removed(test_app):
+    import subprocess
+
+    app = test_app
+
+    srcdir = Path(app.srcdir)
+    out_dir = srcdir / "_build"
+
+    out = subprocess.run(["sphinx-build", "-M", "html", srcdir, out_dir], capture_output=True)
+    assert out.returncode == 2
+
+    assert "Extension error (sphinx_needs.directives.needuml):" in out.stderr.decode("utf-8")
+    assert "exception: 'need' is undefined" in out.stderr.decode("utf-8")